% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBacktesting.R
\name{RBacktestingOutput}
\alias{RBacktestingOutput}
\title{Widget output function for use in Shiny}
\usage{
RBacktestingOutput(outputId, width = "100\%", height = sprintf("\%dpx",
  RCandlesEnv$x$height))
}
\description{
Widget output function for use in Shiny
}

