% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCandles.R
\name{RCandlesOutput}
\alias{RCandlesOutput}
\title{Widget output function for use in Shiny}
\usage{
RCandlesOutput(outputId, width = "100\%", height = sprintf("\%dpx",
  RCandlesEnv$x$height))
}
\description{
Widget output function for use in Shiny
}

