% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBacktesting.R
\name{renderRBacktesting}
\alias{renderRBacktesting}
\title{Widget render function for use in Shiny}
\usage{
renderRBacktesting(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

