% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCandles.R
\name{renderRCandles}
\alias{renderRCandles}
\title{Widget render function for use in Shiny}
\usage{
renderRCandles(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

