% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMotifs.R
\name{renderRMotifs}
\alias{renderRMotifs}
\title{Widget render function for use in Shiny}
\usage{
renderRMotifs(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

