% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMotifs.R
\name{RMotifsOutput}
\alias{RMotifsOutput}
\title{Widget output function for use in Shiny}
\usage{
RMotifsOutput(outputId, width = "100\%", height = sprintf("\%dpx",
  RCandlesEnv$x$height))
}
\description{
Widget output function for use in Shiny
}

